modified

+++ src/main/java/co/taskflow/ecommerce/entity/User.java
@@ -1,5 +1,6 @@
 package co.taskflow.ecommerce.entity;
 
+import co.taskflow.ecommerce.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
@@ -20,7 +21,7 @@ public class User extends BaseEntity implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, length = 50)
     private String username;
